[tcp] adapt the user buffer transmit path to get_writebuf
This pushes much of TCP logic (such as Nagle) further up into the Channel layer. Now, an application queues up a view and writes it immediately, but the app can obtain an entire page to buffer into before doing the write.  We still need to expose congestion status explicitly to the application.